tasks/requirements.txt
celery==5.3.0
redis==5.0.0
python-dotenv==1.0.1
mysql-connector-python==9.1.0